Fix to calculate float

+++ dehazing-sample/include/remove_haze.cuh
@@ -4,8 +4,8 @@
 
 
 __global__ void dehazeGPU(unsigned char* const inputValue, unsigned char* const transmission, unsigned char* dehazedValue,
-                          const double minTransmission, double* const atmosphericLight,
+                          const float minTransmission, float* const atmosphericLight,
                           const int imageWidth, const int imageHeight);
 void executeDehazeGPU(unsigned char* inputValue, unsigned char* const transmission, unsigned char* dehazedValue,
-                      const double minTransmission, double* const atmosphericLight,
+                      const float minTransmission, float* const atmosphericLight,
                       const int imageWidth, const int imageHeight);
